+++ inst/doc/introduction.R
@@ -19,10 +19,13 @@ centers <- data.table(
   x1 = c(5, 0, -3),
   x2 = c(-1, 1, -2)
 )
-points <- centers[, .(
-  x1 = rnorm(num_points, mean = x1),
-  x2 = rnorm(num_points, mean = x2)
-), by = cluster]
+points <- centers[,
+  .(
+    x1 = rnorm(num_points, mean = x1),
+    x2 = rnorm(num_points, mean = x2)
+  ),
+  by = cluster
+]
 
 ggplot(points, aes(x1, x2, color = cluster)) +
   geom_point(alpha = 0.3)
@@ -33,10 +36,10 @@ kclust <- kmeans(points, centers = 3)
 kclust
 
 ## -----------------------------------------------------------------------------
-`%||%` <- function(x, y) if (!is.null(x)) x else y
-
 tidy <- function(x, col.names = colnames(x$centers)) {
-  col.names <- col.names %||% paste0("x", seq_len(ncol(x$centers)))
+  if (is.null(col.names)) {
+    col.names <- paste0("x", seq_len(ncol(x$centers)))
+  }
   dt <- as.data.table(x$centers)
   setnames(dt, col.names)
   dt[, let(
@@ -87,7 +90,8 @@ p1 <- ggplot(assignments, aes(x = x1, y = x2)) +
 p1
 
 ## -----------------------------------------------------------------------------
-p2 <- p1 + geom_point(data = clusters, size = 10, shape = "x") +
+p2 <- p1 +
+  geom_point(data = clusters, size = 10, shape = "x") +
   labs(title = "k-means Clustering with Centers")
 p2